server/requirements.txt
Django==2.2
pytz==2018.9
sqlparse==0.3.0
djangorestframework==3.9.2